requirements.txt
click==8.0.3
Flask==2.0.2
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-Script==2.0.6
Flask-SSLify==0.1.5
is-safe-url==1.0
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
python-dotenv==0.19.2
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.29
Werkzeug==2.0.2
gunicorn==20.1.0
Flask-Moment==1.0.2